\name{NorwayExample}
\docType{data}
\alias{NorwayExample}
\title{Secondary Sampling Unit Data Set}
\description{
 An example data set from Norway containing individuals within postal codes in 1 municipality (Fylke)}
\usage{NorwayExample}
\format{A data set containing 1000 individuals in Oslo, Norway. Each 
  person has a gender/age indicator and are allocated to a postal code region.}
\source{www.ssb.no}
\keyword{ datasets }