\name{NorwayExample_PSU}
\docType{data}
\alias{NorwayExample_PSU}
\title{Primary Sampling Unit Data Set}
\description{
 An example data set from Norway containing population sizes in postal codes}
\usage{NorwayExample_PSU}
\format{A data set containing 944 postal code units for Norway. Each 
  unit is allocated to a municipality (Fylke) and has a population size 
  indicator with gender/age demographic break-down.}
\source{www.ssb.no}
\keyword{ datasets }
