\name{RcmdrPlugin.sampling}
\title{Extra functions to assist with sampling for official Surveys}
\alias{dataonHelp}
\alias{PSUonHelp}
\alias{strataonHelp}
\alias{Stage1onHelp}
\alias{Stage2onHelp}
\alias{mergeDataSets}
\alias{mk.calc}
\alias{special.round}
\alias{disable.fn}
\alias{disable.fn2}
\alias{disable.fn3}
\alias{variableListBox2}
\alias{info.bn}
\alias{setstate}
\alias{setstate2}
\alias{radioButtons2}
\alias{tkblank}
\alias{check.fn}
\alias{check.fn2}
\alias{OKCancelHelp2}
\alias{Sample.datasets}
\alias{PSU.options}
\alias{strata.options}
\alias{SampleSizes}
\alias{SRS.notebook}
\alias{Stage1.notebook}
\alias{Stage2.notebook}
\alias{stage1Sample.fn}
\alias{stage2Sample.fn}
\alias{Strata.notebook}
\alias{StrataSample}
\description{This provides tools to assist with official statistical Surveys, including calculating sample sizes and sampling with a 2-stage design.}
\references{see the EHES manual published on the EHES website:
http://www.ehes.info}
\author{Susie Jentoft and Johan Heldal

Maintainer: susie.jentoft@ssb.no}
\keyword{ survey }
