\name{NorwayExample_Strata}
\docType{data}
\alias{NorwayExample_Strata}
\title{Stratification Data Set}
\description{
 An example data set from Norway containing municipality (Fylke) information 
 including population size, cost of surveying, age variance, and examples of 
  sampling allocation. }
\usage{PSU.df}
\format{A data set containing 19 municipalities for Norway and 11 variables.}
\source{www.ssb.no}
\keyword{ datasets }
