\name{RcmdrPlugin.sampling-package}
\alias{RcmdrPlugin.sampling-package}
\alias{RcmdrPlugin.sampling}
\docType{package}
\title{Sampling for surveys}
\description{This package provides a range of tools to assist with the sampling 
 process for official statistical surveys. Tools include programs to calculate 
 sample sizes and selecting samples in 1- or 2-stages.
 The package is based on the work completed as part of EHES under the package:
 RcdmrPlugin.EHESsampling. The EHES Pilot project has received funding from the 
 European Commission and DG Sanco. The views expressed here are those of the 
 authors and they do not represent Commission's official position.}
\details{
\tabular{ll}{
Package: \tab Rcmdrplugin.sampling\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-23\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{Susie Jentoft and Johan Heldal}
\keyword{ package }
